requirements.txt
langchain==0.1.13
langchain-core==0.1.13
langchain-openai==0.1.1
langchain-cohere==0.1.2
python-dotenv==1.0.1
pinecone-client==3.1.0
pinecone-text
langchain-pinecone==0.0.3
